Add PSRAM double buffering and rewrite arc as triangle fan in C
- Reserve PSRAM front 307KB as graphics back buffer, offset mruby heap
- Add dvi_graphics_set_back_buffer/dvi_graphics_commit for tear-free rendering
- Add dirty flag to skip memcpy when framebuffer is unchanged
- Rewrite fill_arc/draw_arc as triangle fan using sinf/cosf (replaces slow per-pixel angle test)
- Update P5#commit to use DVI::Graphics.commit
- Update dvi_test/p5_test to use DVI::Graphics.commit in wait loops

+++ mrbgems/picoruby-dvi/include/dvi_graphics_draw.h
@@ -87,17 +87,17 @@ void dvi_graphics_draw_ellipse(uint8_t *framebuffer, int width, int height,
                                int cx, int cy, int rx, int ry, uint8_t color);
 
 // Fill an arc (pie slice) centered at (cx, cy) with radius r.
-// Angles are in units of 1/1024 of a full turn (0 = right, 256 = down,
-// 512 = left, 768 = up). This avoids floating-point in the C layer.
+// Angles are in radians (0 = right, PI/2 = down).
+// Rendered as a triangle fan from center using sinf/cosf.
 void dvi_graphics_fill_arc(uint8_t *framebuffer, int width, int height,
                            int cx, int cy, int r,
-                           int start_angle, int stop_angle, uint8_t color);
+                           float start_angle, float stop_angle, uint8_t color);
 
 // Draw an arc outline centered at (cx, cy) with radius r.
-// Angle units are the same as fill_arc (1/1024 of a full turn).
+// Angles are in radians (0 = right, PI/2 = down).
 void dvi_graphics_draw_arc(uint8_t *framebuffer, int width, int height,
                            int cx, int cy, int r,
-                           int start_angle, int stop_angle, uint8_t color);
+                           float start_angle, float stop_angle, uint8_t color);
 
 // Draw a line with a given thickness.
 void dvi_graphics_draw_thick_line(uint8_t *framebuffer, int width, int height,
